add transction api that doesn't require handing back connection
Handing back the connection is quite annoying to work with. This version
is slightly less safe because it will panic code if you go out of your
way to store a reference to the connection outside of the transaction.

However, as evidenced by the test, its pretty convoluted to actually do
this in pratice, so the number of times this actually happens should be
quite small